Rex changed the processing of 4nt to make the equal sign a delimiter like space or tab.  Stupid.
So, after pondering for a bit how to handle it, it seemed
best for me to just change the syntax of a parameter
with a value to a colon instead of an equal sign.

For example, instead of:

xx.btm /a /b=3 /c="parm with space" file1 file2

I changed it to:

xx.btm /a /b:3 /c:"parm with space" file1 file2

Ok, not the fanciest, but it works and I kinda like the
colon a bit better than the equal sign.

--HG--
branch : getopt